Create BookStore.java
I modified some of the data inside the file. Mainly swapping == to .equals because it wasn't properly checking the values. Also implemented a basic version of the menu. I'm not fully sure what you want the printorder function to do in this and don't see a write function after the order(probably because there was no menu and order data available). I didn't comment all the code, but I tried to comment mine to be somewhat helpful to understand what I was doing. Commenting is necessary in the requirements for the phase.