Update BookStore.java
I put the gnumber checker into a separate function and made a function to make sure all the strings existed and used them on the createaccount and login functions of the code. This is so we had try catch exception handling we didn't have previously. I hope this change doesn't impact your code too much.